Artemis: mc: Support to update cxl firmware
Summary:
- Support PM8702 firmware update.
- Modify BIC i2c1 frequency to 100K to match PM8702 setting.
- Correct component id for geting firmware version.
- Support ipmi command to activate PM8702 firmware after firmware successfully updated.

Test Plan:
- Build code: Pass
- Update PM8702 firmware: Pass
- Get cxl dimm temperature: Pass

Log:
- Update PM8702 firmware

root@bmc-oob:~# fw-util meb --update cxl7 Artemis_G4_v61.15.00.00_1150-fwdl.bin
updating fw on bus 9
updated: 100 %
Update success, sleep 10 secs and active cxl7

Elapsed time:  669   sec.
Upgrade of meb : cxl7 succeeded
root@bmc-oob:~# fw-util meb --version cxl7
MEB CXL7 Version: 01.002.003.00

root@bmc-oob:~# power-util sled-cycle
packet_write_wait: Connection to fe80::e63d:1aff:fe34:8819%enp7s0 port 22: Broken pipe
lora@guaigaui-Freedom:~/zephyrproject/yv3.5_zephyr_bic$ ssh root@fe80::e63d:1aff:fe34:8819%enp7s0
root@fe80::e63d:1aff:fe34:8819%enp7s0's password:
Last login: Fri Mar  9 04:36:07 2018 from fe80::7efe:90ff:fe58:76b2%eth0
root@bmc-oob:~# fw-util meb --version cxl7
MEB CXL7 Version: 01.001.005.00

- Get cxl dimm temperature

root@bmc-oob:~# sensor-util all | grep DIMM
......
MEB_CXL8_DIMM_A_TEMP_C       (0x21) :   26.50 C     | (ok)
MEB_CXL8_DIMM_B_TEMP_C       (0x22) :   26.75 C     | (ok)
MEB_CXL8_DIMM_C_TEMP_C       (0x23) :   26.25 C     | (ok)
MEB_CXL8_DIMM_D_TEMP_C       (0x24) :   26.75 C     | (ok)
MEB_CXL7_DIMM_A_TEMP_C       (0x21) :   24.25 C     | (ok)
MEB_CXL7_DIMM_B_TEMP_C       (0x22) :   24.75 C     | (ok)
MEB_CXL7_DIMM_C_TEMP_C       (0x23) :   24.25 C     | (ok)
MEB_CXL7_DIMM_D_TEMP_C       (0x24) :   24.50 C     | (ok)
MEB_CXL6_DIMM_A_TEMP_C       (0x21) :   26.25 C     | (ok)
MEB_CXL6_DIMM_B_TEMP_C       (0x22) :   26.75 C     | (ok)
MEB_CXL6_DIMM_C_TEMP_C       (0x23) :   26.50 C     | (ok)
MEB_CXL6_DIMM_D_TEMP_C       (0x24) :   26.50 C     | (ok)
MEB_CXL5_DIMM_A_TEMP_C       (0x21) :   26.00 C     | (ok)
MEB_CXL5_DIMM_B_TEMP_C       (0x22) :   26.50 C     | (ok)
MEB_CXL5_DIMM_C_TEMP_C       (0x23) :   26.00 C     | (ok)
MEB_CXL5_DIMM_D_TEMP_C       (0x24) :   26.75 C     | (ok)
MEB_CXL3_DIMM_A_TEMP_C       (0x21) :   26.50 C     | (ok)
MEB_CXL3_DIMM_B_TEMP_C       (0x22) :   27.00 C     | (ok)
MEB_CXL3_DIMM_C_TEMP_C       (0x23) :   27.00 C     | (ok)
MEB_CXL3_DIMM_D_TEMP_C       (0x24) :   27.25 C     | (ok)
MEB_CXL4_DIMM_A_TEMP_C       (0x21) :   26.00 C     | (ok)
MEB_CXL4_DIMM_B_TEMP_C       (0x22) :   26.25 C     | (ok)
MEB_CXL4_DIMM_C_TEMP_C       (0x23) :   26.25 C     | (ok)
MEB_CXL4_DIMM_D_TEMP_C       (0x24) :   26.25 C     | (ok)
MEB_CXL1_DIMM_A_TEMP_C       (0x21) :   27.00 C     | (ok)
MEB_CXL1_DIMM_B_TEMP_C       (0x22) :   27.25 C     | (ok)
MEB_CXL1_DIMM_C_TEMP_C       (0x23) :   27.25 C     | (ok)
MEB_CXL1_DIMM_D_TEMP_C       (0x24) :   27.50 C     | (ok)
MEB_CXL2_DIMM_A_TEMP_C       (0x21) :   27.25 C     | (ok)
MEB_CXL2_DIMM_B_TEMP_C       (0x22) :   27.50 C     | (ok)
MEB_CXL2_DIMM_C_TEMP_C       (0x23) :   27.25 C     | (ok)
MEB_CXL2_DIMM_D_TEMP_C       (0x24) :   27.75 C     | (ok)

+++ common/dev/include/pm8702.h
@@ -81,8 +81,7 @@ typedef enum _pm8702_access {
 bool pm8702_get_dimm_temp(void *mctp_p, mctp_ext_params ext_params, uint16_t address,
 			  int16_t *interger, int16_t *fraction);
 bool pm8702_cmd_handler(void *mctp_inst, mctp_ext_params ext_params, uint16_t opcode,
-			uint8_t *data_buf, uint8_t data_len, uint8_t *response,
-			uint8_t *response_len);
+			uint8_t *data_buf, int data_len, uint8_t *response, uint8_t *response_len);
 
 #endif
